plugin auto2000 test is split up to four tests
