+++ test/plugin_auto2000/PluginAuto2000Tests.cpp
@@ -1,0 +1,5 @@
+#include "PluginAuto2000Tests.h"
+
+PluginAuto2000Tests::PluginAuto2000Tests() {
+    pluginsModelsDir = rrTestModelsDir_ / "PLUGINS";
+}
